feat: add CodeMirror-based code viewer for repository files (#136)
Added codemirror, so that when you click on a file in the file tree, it opens the editor as a tab. I went with codemirror since it seems a bit more flexible than monaco, but we can swap this out easily if needed.

Will add syntax highlighting, diffs, etc. in other PRs.

![image.png](https://app.graphite.com/user-attachments/assets/ce3e0388-a05c-4637-b672-2866bf31bb6a.png)