feat: make terminals persistent and serializable
We now persist shells when the ShellTerminal is unmounted/remounted in React, this means it'll keep running in the background like you're used to in other editors. It also serializes so that it retains terminal state on restart. CMD+K to clear now also works properly.
After the monorepo move, node-pty wasn't being rebuilt, this also fixes that.