feat: temporary tabs
Now when you click a tab in the editor, it'll open as a temporary tab.
Double clicking the file again (either in the changelist, filetree, or the tab itself), makes it a permanent one.
Also fixed an issue where navigating through the ChangesPanel would open a new tab every time.