feat: Initial task deep linking (#283)
The "create if not exists locally" functionality needs work, unsure how to infer repoPath if it's never been setup. For now, it will open a task if you deep link to it, or if the app is opened for the first time by deep link if the task/workspace already existing locally.